Update legal.md
Hi,

I just noticed you translated the page into Chinese and published it here: 

http://cheats.rs.kingfree.moe/

First of all, wow, that must have been quite a bit of work. I'm really happy to see that you made the page accessible to a wider audience!

One small thing though, your published legal page here:

http://cheats.rs.kingfree.moe/legal/

says I'm the "Operator", which is the site owner legally responsible for the hosting on that site. Unfortunately this is not true. I would appreciate if you could merge this PR (or just change the text yourself) and re-publish an update removing my name from that line.

Thanks,
Ralf